pyproject.toml
[project]
name = "adv-archon"
version = "0.2.0"
description = "Asistente IA personal 100% local — motor Ollama, acceso al sistema con consentimiento"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "chromadb>=0.5.0",
    "rich>=13.7.0",
    "prompt_toolkit>=3.0.43",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pathspec>=0.12.0",
    "typer>=0.12.0",
    "watchdog>=4.0.0",
]

[project.optional-dependencies]
voice = [
    "faster-whisper>=1.0.0",
    "pyttsx3>=2.90",
]

[project.scripts]
archon = "adv_archon.__main__:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
